refactor: place utility methods in pkg/cmd
