+++ pkg/cmd/kafka/consumergroup/describe/describe.go
@@ -7,8 +7,8 @@ import (
 	"net/http"
 	"sort"
 
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/kafka/consumergroup/consumergrouputil"
 	kafkacmdutil "github.com/redhat-developer/app-services-cli/pkg/kafkautil"
-	"github.com/redhat-developer/app-services-cli/pkg/kafkautil/consumergrouputil"
 
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
