+++ pkg/cmd/kafka/shared/kafka_util_test.go
@@ -1,4 +1,4 @@
-package kafkautil
+package shared
 
 import (
 	"testing"
