+++ pkg/cmd/kafka/topic/list/list.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/redhat-developer/app-services-cli/pkg/kafkautil/topicutil"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/kafka/topic/topicutil"
 
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
