+++ pkg/cmd/kafka/topic/topicutil/util.go
