+++ pkg/cmd/registry/artifact/crud/list/list.go
@@ -2,8 +2,9 @@ package list
 
 import (
 	"context"
+
 	"github.com/redhat-developer/app-services-cli/pkg/cmd/registry/artifact/util"
-	"github.com/redhat-developer/app-services-cli/pkg/serviceregistryutil"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/registry/registryutil"
 
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
@@ -135,7 +136,7 @@ func runList(opts *options) error {
 
 	response, _, err := request.Execute()
 	if err != nil {
-		return serviceregistryutil.TransformInstanceError(err)
+		return registryutil.TransformInstanceError(err)
 	}
 
 	if len(response.Artifacts) == 0 && opts.outputFormat == "" {
