+++ pkg/cmd/registry/artifact/role/list/list.go
@@ -2,8 +2,8 @@ package list
 
 import (
 	"context"
-	"github.com/redhat-developer/app-services-cli/pkg/serviceregistryutil"
 
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/registry/registryutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/config"
@@ -100,7 +100,7 @@ func runList(opts *options) error {
 	}
 	mappings, _, err := a.AdminApi.ListRoleMappings(opts.Context).Execute()
 	if err != nil {
-		return serviceregistryutil.TransformInstanceError(err)
+		return registryutil.TransformInstanceError(err)
 	}
 
 	if len(mappings) == 0 && opts.outputFormat == "" {
