+++ pkg/cmd/registry/artifact/role/revoke/revoke.go
@@ -2,13 +2,14 @@ package revoke
 
 import (
 	"context"
+
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/registry/registryutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
 	"github.com/redhat-developer/app-services-cli/pkg/core/config"
 	"github.com/redhat-developer/app-services-cli/pkg/core/connection"
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/iostreams"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"github.com/redhat-developer/app-services-cli/pkg/core/logging"
-	"github.com/redhat-developer/app-services-cli/pkg/serviceregistryutil"
 	"github.com/spf13/cobra"
 )
 
@@ -97,7 +98,7 @@ func runRevoke(opts *options) error {
 
 	_, err = dataAPI.AdminApi.DeleteRoleMapping(opts.Context, opts.principal).Execute()
 	if err != nil {
-		return serviceregistryutil.TransformInstanceError(err)
+		return registryutil.TransformInstanceError(err)
 	}
 	opts.Logger.Info(opts.localizer.MustLocalize("registry.role.cmd.revoke.success"))
