+++ pkg/cmd/registry/registryutil/registry_util.go
@@ -1,4 +1,4 @@
-package serviceregistryutil
+package registryutil
 
 import (
 	"errors"
