+++ pkg/cmd/serviceaccount/delete/delete.go
@@ -2,6 +2,9 @@ package delete
 
 import (
 	"context"
+	"net/http"
+
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/serviceaccount/shared/validation"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/config"
@@ -10,8 +13,6 @@ import (
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/iostreams"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"github.com/redhat-developer/app-services-cli/pkg/core/logging"
-	"github.com/redhat-developer/app-services-cli/pkg/serviceaccountutil/validation"
-	"net/http"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/spf13/cobra"
